[wasm] Fix WebCIL segment test global index type
DwarfHelper.ReadULEB128 returns ulong, so keep the parsed active payload global index as ulong before comparing it with the WebCIL image base global index.


Copilot-Session: a519827b-7ffd-4499-92f6-8f94d61ea940

+++ src/coreclr/tools/aot/ILCompiler.ReadyToRun.Tests/TestCases/R2RTestSuites.cs
@@ -2292,7 +2292,7 @@ public class R2RTestSuites
                 Assert.True(payloadSegmentKind == 0, "Expected second segment to be active (kind 0)");
                 int payloadSegmentSizeOffset = payloadSegmentOffset + 1;
                 Assert.True(imageSpan[payloadSegmentSizeOffset] == 0x23, "Expected active payload segment offset to use global.get");
-                uint payloadSegmentOffsetGlobalIndex = DwarfHelper.ReadULEB128(imageSpan.Slice(payloadSegmentSizeOffset + 1), out int globalIndexBytes);
+                ulong payloadSegmentOffsetGlobalIndex = DwarfHelper.ReadULEB128(imageSpan.Slice(payloadSegmentSizeOffset + 1), out int globalIndexBytes);
                 Assert.True(payloadSegmentOffsetGlobalIndex == WebCilObjectWriter.ImageBaseGlobalIndex,
                     $"Expected active payload segment offset to use image base global {WebCilObjectWriter.ImageBaseGlobalIndex}, but got {payloadSegmentOffsetGlobalIndex}");
                 payloadSegmentSizeOffset += 1 + globalIndexBytes;
